revert: return useEffect and useContext usage from availability-body.tsx to stable version
This reverts commit 7d82726ff8bb00de57c3b69033fc90014787223b.

tttttt branch is up to date with 'origin/169-refactor-usecontext-and-useeffect-usage-on-availability-pages'.
 revert: return to stable version#	modified:   src/components/availability/personal-availability.tsx
